handles new edge cases
In this version, instead of sorting into buckets, we split into subarrays. Earlier, we took the file line by line, and put each line into one of four piles, timestamp, patient, interviewer, or discard. Here, we take a file and remove anything that is not a time or a dialogue. Then we split it into two arrays of subarrays. The first contains times, the second contains augmented pairs. That is, if during a timestamp, the interviewer (or patient) has spoken but the other person hasn't we augment that with the word 'blank'. We use this to build the dataframe